refactor(aides): move Aides-Territoires to Grist logos init
